update structure of the project
