+++ README.md
@@ -1,8 +1,8 @@
-# Culqi-Java
-Cliente en Java del API de Culqi.
+# com.culqi.Culqi-Java
+Cliente en Java del API de com.culqi.Culqi.
 
 # Licencia
 
-Copyright 2016 Culqi
+Copyright 2016 com.culqi.Culqi
 
 Licensed under the MIT License
