+++ src/main/java/com/culqi/controller/Charges.java
@@ -1,31 +1,31 @@
-package controller;
+package com.culqi.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import model.Refund;
-import model.Secure;
-import modelreponse.RefundResponse;
+import com.culqi.model.Charge;
+import com.culqi.model.Secure;
+import com.culqi.modelreponse.ChargeResponse;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
 import org.apache.http.util.EntityUtils;
-import util.Result;
-import util.Util;
+import com.culqi.util.Result;
+import com.culqi.util.Util;
 
 /**
- * Created by culqi on 12/23/16.
+ * Created by culqi on 12/22/16.
  */
-public class Refunds {
+public class Charges {
 
-    private static final String URL = "/refunds/";
+    private static final String URL = "/charges/";
 
     Util util = new Util();
 
     ObjectMapper mapper = new ObjectMapper();
 
-    public Result createRefund(Secure secure, Refund refund) throws Exception {
+    public Result createCharge(Secure secure, Charge charge) throws Exception {
         Result result = new Result();
         result.setMessage("CODE STATUS NOT SUPPORTED");
         HttpResponse response;
-        String jsonData = mapper.writeValueAsString(refund);
+        String jsonData = mapper.writeValueAsString(charge);
         response = util.response(secure, URL, jsonData);
         HttpEntity entity = response.getEntity();
         String statusCode = response.getStatusLine().toString();
@@ -37,11 +37,11 @@ public class Refunds {
             result.setMessage(errorMessage);
         }
         if(statusCode.contains("201")) {
-            RefundResponse refundResponse = mapper.readValue(jsonResult, RefundResponse.class);
-            result.setId(refundResponse.getId());
-            result.setMessage(refundResponse.getReason());
+            ChargeResponse chargeResponse = mapper.readValue(jsonResult, ChargeResponse.class);
+            result.setId(chargeResponse.getId());
+            result.setMessage(chargeResponse.getMerchant_message());
         }
-        return  result;
+        return result;
     }
 
 }
