+++ src/main/java/com/culqi/controller/Plans.java
@@ -1,31 +1,31 @@
-package controller;
+package com.culqi.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import model.Charge;
-import model.Secure;
-import modelreponse.ChargeResponse;
+import com.culqi.model.Plan;
+import com.culqi.model.Secure;
+import com.culqi.modelreponse.PlanResponse;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
 import org.apache.http.util.EntityUtils;
-import util.Result;
-import util.Util;
+import com.culqi.util.Result;
+import com.culqi.util.Util;
 
 /**
  * Created by culqi on 12/22/16.
  */
-public class Charges {
+public class Plans {
 
-    private static final String URL = "/charges/";
+    private static final String URL = "/plans/";
 
     Util util = new Util();
 
     ObjectMapper mapper = new ObjectMapper();
 
-    public Result createCharge(Secure secure, Charge charge) throws Exception {
+    public Result createPlan(Secure secure, Plan plan) throws Exception {
         Result result = new Result();
         result.setMessage("CODE STATUS NOT SUPPORTED");
         HttpResponse response;
-        String jsonData = mapper.writeValueAsString(charge);
+        String jsonData = mapper.writeValueAsString(plan);
         response = util.response(secure, URL, jsonData);
         HttpEntity entity = response.getEntity();
         String statusCode = response.getStatusLine().toString();
@@ -37,9 +37,9 @@ public class Charges {
             result.setMessage(errorMessage);
         }
         if(statusCode.contains("201")) {
-            ChargeResponse chargeResponse = mapper.readValue(jsonResult, ChargeResponse.class);
-            result.setId(chargeResponse.getId());
-            result.setMessage(chargeResponse.getMerchant_message());
+            PlanResponse planResponse = mapper.readValue(jsonResult, PlanResponse.class);
+            result.setId(planResponse.getId());
+            result.setMessage(planResponse.getAlias());
         }
         return result;
     }
