+++ src/main/java/com/culqi/controller/Refunds.java
@@ -1,31 +1,31 @@
-package controller;
+package com.culqi.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import model.Plan;
-import model.Secure;
-import modelreponse.PlanResponse;
+import com.culqi.model.Refund;
+import com.culqi.model.Secure;
+import com.culqi.modelreponse.RefundResponse;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
 import org.apache.http.util.EntityUtils;
-import util.Result;
-import util.Util;
+import com.culqi.util.Result;
+import com.culqi.util.Util;
 
 /**
- * Created by culqi on 12/22/16.
+ * Created by culqi on 12/23/16.
  */
-public class Plans {
+public class Refunds {
 
-    private static final String URL = "/plans/";
+    private static final String URL = "/refunds/";
 
     Util util = new Util();
 
     ObjectMapper mapper = new ObjectMapper();
 
-    public Result createPlan(Secure secure, Plan plan) throws Exception {
+    public Result createRefund(Secure secure, Refund refund) throws Exception {
         Result result = new Result();
         result.setMessage("CODE STATUS NOT SUPPORTED");
         HttpResponse response;
-        String jsonData = mapper.writeValueAsString(plan);
+        String jsonData = mapper.writeValueAsString(refund);
         response = util.response(secure, URL, jsonData);
         HttpEntity entity = response.getEntity();
         String statusCode = response.getStatusLine().toString();
@@ -37,11 +37,11 @@ public class Plans {
             result.setMessage(errorMessage);
         }
         if(statusCode.contains("201")) {
-            PlanResponse planResponse = mapper.readValue(jsonResult, PlanResponse.class);
-            result.setId(planResponse.getId());
-            result.setMessage(planResponse.getAlias());
+            RefundResponse refundResponse = mapper.readValue(jsonResult, RefundResponse.class);
+            result.setId(refundResponse.getId());
+            result.setMessage(refundResponse.getReason());
         }
-        return result;
+        return  result;
     }
 
 }
