+++ src/main/java/com/culqi/model/CardHolder.java
@@ -1,4 +1,4 @@
-package model;
+package com.culqi.model;
 
 import lombok.Data;
