+++ src/main/java/com/culqi/model/Client.java
@@ -1,4 +1,4 @@
-package model;
+package com.culqi.model;
 
 import lombok.Data;
