+++ src/main/java/com/culqi/modelreponse/ErrorResponse.java
@@ -1,4 +1,4 @@
-package modelreponse;
+package com.culqi.modelreponse;
 
 import lombok.Data;
