+++ src/main/java/com/culqi/modelreponse/RefundResponse.java
@@ -1,4 +1,4 @@
-package modelreponse;
+package com.culqi.modelreponse;
 
 import lombok.Data;
