+++ src/main/java/com/culqi/util/Result.java
@@ -1,4 +1,4 @@
-package util;
+package com.culqi.util;
 
 import lombok.Data;
